pyproject.toml
[project]
name = "agentes-de-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "sentence-transformers>=5.3.0",
]
